+++ src/core/event_loop.rs
@@ -36,7 +36,7 @@ use crate::{
             run_status_update,
         },
     },
-    ui::{KeyHandleResult, Ui, UiEvent},
+    ui::{KeyHandleResult, Ui, UiAppEvent, UiEvent, modals::info_modal::InfoModal},
 };
 
 static ON_RESIZE_SCHEDULE_ID: LazyLock<Id> = LazyLock::new(id::new);
@@ -190,6 +190,21 @@ fn main_task<B: Backend + std::io::Write>(
                     context.scheduler.schedule(Duration::from_secs(5), |(tx, _)| {
                         Ok(tx.send(AppEvent::RequestRender)?)
                     });
+                }
+                AppEvent::InfoModal { message, title, size } => {
+                    if let Err(err) = ui.on_ui_app_event(
+                        UiAppEvent::Modal(Box::new(
+                            InfoModal::builder()
+                                .context(&context)
+                                .maybe_title(title)
+                                .maybe_size(size)
+                                .message(message)
+                                .build(),
+                        )),
+                        &mut context,
+                    ) {
+                        log::error!(error:? = err; "UI failed to handle modal event");
+                    }
                 }
                 AppEvent::Log(msg) => {
                     if let Err(err) = ui.on_event(UiEvent::LogAdded(msg), &mut context) {